// app/(apps)/revoke/types.ts
export type TAllowance = {
	address: TAddress;
	args: {
		owner: TAddress;
		sender: TAddress;
		value: string | number | bigint | undefined;
	};
	blockNumber: bigint;
	chainID: number;
	logIndex: number;
	transactionIndex: number;
};
// ...
export type TExpandedAllowance = TAllowance & {
	name: string;
	symbol: string;
	decimals: number;
	balanceOf: TNormalizedBN;
	spenderName: string;
};
// ...
export type TFilterAllowance = Pick<TExpandedAllowance, 'symbol' | 'chainID' | 'address' | 'args'> & {
	displayName?: TAddress | string;
	spenderName: string;
};